requirements.txt
flask
requests
Pillow
colorthief
gunicorn